requirements.txt
Django>=3.1,<3.2
djangorestframework >=3.11.1
